Fix guests not showing up in class user list
